db-api: Remove default for Language component schema
There's probably a better way to do this with selective usage of
Parameter() and schema_component_key, but I'm too short on time to
figure out how to do this neatly at the moment.

I think this change only works because the staff schemas are processed
after the novel schemas. If language is used elsewhere (it will be),
we'll probably need to refactor this to make it more clear and not
flaky.