db-api: Rename retrieve_user_handler
I don't think we can return an ORM AuthUser object at this point because
the metadata stuff hasn't been set up. Using .scalar() will just return
the first column in the row (the user_id). But we already knew that from
the Litestar session. So rename it to retrieve_user_id_handler to try to
make it more understandable.